+++ src/Module/Oxideshop.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Copyright © OXID eSales AG. All rights reserved.
  * See LICENSE file for license details.
@@ -6,17 +7,17 @@
 
 namespace OxidEsales\Codeception\Module;
 
-require_once __DIR__.'/../../../../oxid-esales/testing-library/base.php';
+require_once __DIR__ . '/../../../../oxid-esales/testing-library/base.php';
 
+use Codeception\Exception\ElementNotFound;
 use Codeception\Lib\Interfaces\DependsOnModule;
-use Codeception\Module\WebDriver;
+use Codeception\Module;
 use Codeception\Module\Db;
+use Codeception\Module\WebDriver;
+use Facebook\WebDriver\Exception\ElementNotVisibleException;
+use Facebook\WebDriver\Exception\NoSuchElementException;
 
-/**
- * Class Oxideshop
- * @package OxidEsales\Codeception\Module
- */
-class Oxideshop extends \Codeception\Module implements DependsOnModule
+class Oxideshop extends Module implements DependsOnModule
 {
     /**
      * @var WebDriver
@@ -152,5 +153,26 @@ class Oxideshop extends \Codeception\Module implements DependsOnModule
     public function seePageHasElement($element)
     {
         return count($this->getModule('WebDriver')->_findElements($element)) > 0;
+    }
+
+    /**
+     * Clicks on first visible element
+     * @param string $locator
+     * @throws ElementNotVisibleException
+     * @throws NoSuchElementException
+     */
+    public function seeAndClick(string $locator): void
+    {
+        $elements = $this->webDriver->_findElements($locator);
+        if (!$elements) {
+            throw new NoSuchElementException($locator);
+        }
+        foreach ($elements as $el) {
+            if ($el->isDisplayed()) {
+                $el->click();
+                return;
+            }
+        }
+        throw new ElementNotVisibleException($locator);
     }
 }