+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Flow theme module
+- Methods:
+    - `Module\Oxideshop::seeAndClick`
 
 ## [1.2.0] - 2020-01-02